WIP: streamer example showing message streaming.
A C example, similar to broker.c but only allows a single sender/receiver per
address. Sending 10x4Mb messages, the streamer stays below about 1.5M in size
where the broker example grows to about 25Mb.

Performance of the streamer is horrible - partly because it has unreasonably
small buffers to test that it is correct when messages are much bigger than
buffer space. However we should make sure there aren't other problems. Dispatch
has a more sophisticated buffering scheme so this example may not be realistic
in that regard.